Updated the FileTools class.
Made it so it stores the username of the user and associates it with the ID, also adjusted the functions to incorporate the username.

Added the ISA as something to get stored also. So now by default the value stored is 'n' for null, and it should have verification implemented in the ISA class making it so only 'c', 'f', 'l' can be stored.